requirements.txt
Flask==2.2.2
python-dotenv
Flask-DebugToolbar
Flask-SQLAlchemy
Flask-Login
cryptography
wtforms[email]
Pillow